Version 6. Add inteferences (with Alpha distributions)

+++ src/flying_object/FlyingGroup.java
@@ -263,7 +263,7 @@ public class FlyingGroup extends MobileG {
 	}
 	
 	@Override
-	public long getNextTime() { return 0 ;}
+	public double getNextTime() { return 0 ;}
 	
 	@Override
 	public void loadRouteFromFile() {}
