+++ src/natural_events/Gas.java
@@ -185,7 +185,7 @@ public class Gas extends DeviceWithoutRadio {
 	}
 	
 	@Override
-	public long getNextTime() { return 0 ;}
+	public double getNextTime() { return 0 ;}
 	
 	@Override
 	public void loadRouteFromFile() {}
