+++ src/script/Command_PICK.java
@@ -12,12 +12,12 @@ public class Command_PICK extends Command {
 	}
 
 	@Override
-	public long execute() {
+	public double execute() {
 //		int lEvent = sensor.pickMessage(arg);
 //		double ratio = (DataInfo.ChDataRate*1.0)/(DataInfo.UartDataRate);
 //		return (int)(Math.round(lEvent*8.*ratio));
 		sensor.pickMessage(arg);
-		return 0;
+		return 0 ;
 	}
 
 	@Override
